Add waitress serving and bump to 1.0.0 for first production ready release (#1)
waitress.serve() offers a multi-threaded solution for hosting a dash app.

TAA filtered probability maps were removed as they are out of scope for this initial release.

Version number has been bumped to the first major release, as I believe this version of the code is suitable for deployment on the server.